requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.7.22
charset-normalizer==3.2.0
Django==4.2.3
idna==3.4
requests==2.31.0
soupsieve==2.4.1
sqlparse==0.4.4
typing_extensions==4.7.1
urllib3==2.0.4
whitenoise==6.5.0
python-dotenv==1.0.0